% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_items.R
\docType{package}
\name{packup-package}
\alias{packup}
\alias{packup-package}
\title{packup: Greedy Heap-Based Weight Distributor}
\description{
Tool to distibute weight among people using heap-based priority queue implemented in C.
}
